Add files via upload
Complete the getMonthlyPrice function. It accepts a tier (string) as input and returns the monthly price for that tier in pennies. Here are the prices in dollars:

"basic" - $100.00
"premium" - $150.00
"enterprise" - $500.00
Convert the prices from dollars to pennies. If the given tier doesn't match any of the above, return 0 pennies.

To avoid pesky floating-point errors, we often store prices in the currency's base unit. In this case, we are storing the prices in pennies, and a dollar consists of 100 pennies.